django/requirements.txt
Django==1.11b1
djangorestframework==3.6.1
psycopg2==2.7